Escuela ABC Completa

+++ maestros/routes.py
@@ -1,8 +1,7 @@
-from flask import render_template, request, redirect, url_for
+from flask import render_template, request, redirect, url_for, flash
 from . import maestros
 from models import db, Maestros
 import forms
-
 
 @maestros.route("/maestros", methods=['GET'])
 def index():
@@ -14,7 +13,9 @@ def index():
 @maestros.route("/maestros_agregar", methods=['GET', 'POST'])
 def agregar():
     create_form = forms.MaestroForm(request.form)
-    if request.method == 'POST':
+    
+    
+    if create_form.validate_on_submit():
         maestro_nuevo = Maestros(
             nombre=create_form.nombre.data,
             apellidos=create_form.apellidos.data,
@@ -25,6 +26,7 @@ def agregar():
         db.session.commit()
         return redirect(url_for('maestros.index'))
     
+    
     return render_template("maestros/agregar.html", form=create_form)
 
 
@@ -32,31 +34,26 @@ def agregar():
 def detalles():
     create_form = forms.MaestroForm(request.form)
     matricula = request.args.get('id')
-    maestro = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
+    maestro = Maestros.query.get(matricula)
     
     return render_template('maestros/detalles.html', maestro=maestro, form=create_form)
 
 
 @maestros.route("/maestros_modificar", methods=['GET', 'POST'])
 def modificar():
-    create_form = forms.MaestroForm(request.form)
+    matricula = request.args.get('id')
+    maestro = Maestros.query.get(matricula)
     
-    if request.method == 'GET':
-        matricula = request.args.get('id')
-        maestro = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
-        create_form.matricula.data = maestro.matricula
-        create_form.nombre.data = maestro.nombre
-        create_form.apellidos.data = maestro.apellidos
-        create_form.email.data = maestro.email
-        create_form.especialidad.data = maestro.especialidad
-
-    if request.method == 'POST':
-        matricula = request.args.get('id')
-        maestro = db.session.query(Maestros).filter(Maestros.matricula == matricula).first()
+    
+    create_form = forms.MaestroForm(request.form, obj=maestro)
+    
+    
+    if create_form.validate_on_submit():
         maestro.nombre = create_form.nombre.data
         maestro.apellidos = create_form.apellidos.data
         maestro.email = create_form.email.data
         maestro.especialidad = create_form.especialidad.data
+        
         db.session.commit()
         return redirect(url_for('maestros.index'))
 
@@ -79,6 +76,12 @@ def eliminar():
     if request.method == 'POST':
         matricula = create_form.matricula.data
         maestro = Maestros.query.get(matricula)
+        
+       
+        if maestro.cursos:
+            flash("No se puede eliminar: El maestro tiene cursos asignados.")
+            return redirect(url_for('maestros.index'))
+
         db.session.delete(maestro)
         db.session.commit()
         return redirect(url_for('maestros.index'))
